exporter/requirements.txt
boto3==1.34.69
prometheus_client==0.21.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
